requirements.txt
Flask==0.12.2
numpy
scipy
scikit-learn
sklearn-crfsuite
nltk